requirements.txt
tensorflow
nvidia-pyindex
nvidia-tensorflow
gradio
fastapi
uvicorn
pillow
kaggle
matplotlib